Split BlockHouseHolder_DDRB into InnerHouseholder + Householder homes (Phase 1.2a)
Distributes the block-QR helper file into two shared, neutrally-packaged homes
(org.ejml.dense.block) so QR, Tridiagonal, and future Bidiagonal draw their
Householder primitives from one place. No merging; method bodies move verbatim with
their //CONCURRENT hints.

- InnerHouseholder_DDRB (L2, ~xLARFG + xLARF): the unblocked per-reflector ops
  (computeHouseholderCol/Row, rank1UpdateMultR_Col/_TopRow, rank1UpdateMultL_Row/_LeftCol,
  innerProdCol/Row, findMaxCol/Row, computeTauAndDivideCol/Row, divideElementsCol,
  scale_row, add_row).
- Householder_DDRB (L3, ~xLARFT + xLARFB): the blocked WY ops (computeW_Column,
  initializeW, computeZ, computeY_t_V, multAdd_zeros, multTransA_vecCol). The raw
  zero/one kernels multBlockAdd_zerosone and multTransABlockSet_lowerTriag become
  private members (Householder-specific, not generic block multiply).
- decomposeQR_block_col (the panel algorithm) moves into QRDecompositionHouseholder_DDRB
  and calls the L2 ops; kept public/static (tests exercise it directly).
- MatrixOps_DDRB.lastBlockStart(n, blockLength) extracts the shared "start index of the
  last reflector block"; QR.applyQ repointed to it.
- Rewired QR //CONCURRENT_MACRO (InnerHouseholder/Householder), Tridiagonal/Bidiagonal
  helper call sites; removed dead commented lines in QR.applyQTran; deleted BlockHouseHolder_DDRB.
- Tests split to mirror the new homes: TestInnerHouseholder_DDRB (+ _MT) and
  TestHouseholder_DDRB (+ _MT); decomposeQR_block_col test moved into the QR test (+ _MT);
  TestBlockHouseHolder_DDRB (+ _MT) removed.

Serial + _MT block decomposition/solver tests green (ddense + fdense).
